Migrate: to continuous basic/combat/health/mana stats. Part 2

+++ src/functions/calcDamage.js
@@ -1,5 +1,5 @@
 export default function calcDamage(PAtk, PDef) {
   const k = Number(PAtk) / Number(PDef)
-  let d = BigInt(Math.trunc(Number(PAtk) * k))
-  return d > 0n ? d : 1n
+  let d = Math.trunc(Number(PAtk) * k)
+  return d > 0 ? d : 1
 }
